# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TG Restaurant ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List all branches
      responses:
        "200":
          description: Branch list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    patch:
      operationId: updateBranch
      tags: [branches]
      summary: Update branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── USERS ────────────────────────────────────────────────
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Super admin command center summary
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
  /dashboard/alerts:
    get:
      operationId: getDashboardAlerts
      tags: [dashboard]
      summary: Alerts (low stock, delayed orders, failed deliveries)
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /dashboard/top-items:
    get:
      operationId: getTopMenuItems
      tags: [dashboard]
      summary: Top selling menu items
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopMenuItem"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # BRANCH
    Branch:
      type: object
      required: [id, name, address, phone, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
    BranchInput:
      type: object
      required: [name, address, phone]
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        active:
          type: boolean
    BranchUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        active:
          type: boolean

    # USER
    User:
      type: object
      required: [id, name, phone, role, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, branch_manager, kitchen_staff, delivery_staff, order_staff, addis_staff]
        branchId:
          type: ["integer", "null"]
        baseSalary:
          type: ["number", "null"]
        active:
          type: boolean
        createdAt:
          type: string
...
    DashboardSummary:
      type: object
      required: [todayOrders, todayRevenue, activeDeliveries, pendingKitchen, lowStockAlerts]
      properties:
        todayOrders:
          type: integer
        todayRevenue:
          type: number
        activeDeliveries:
          type: integer
        pendingKitchen:
          type: integer
        lowStockAlerts:
          type: integer
        avgPrepMinutes:
          type: ["number", "null"]
        avgDeliveryMinutes:
          type: ["number", "null"]
        lotteryDrawToday:
          type: boolean
        onlineStaff:
          type: integer
    Alert:
      type: object
      required: [id, type, severity, message, branchId, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [low_stock, delayed_order, failed_delivery, system]
        severity:
          type: string
          enum: [info, warning, critical]
        message:
          type: string
        branchId:
          type: integer
        branchName:
          type: ["string", "null"]
        createdAt:
          type: string
    TopMenuItem:
      type: object
      required: [menuItemId, nameEn, totalSold, totalRevenue]
      properties:
        menuItemId:
          type: integer
        nameEn:
          type: string
        nameAm:
          type: ["string", "null"]
        totalSold:
          type: integer
        totalRevenue:
          type: number
        photoUrl:
          type: ["string", "null"]
...